LayoutTests/accessibility/selection-states.html

<html>
<head>
    <script src="../resources/js-test.js"></script>
</head>

<body id="body">

<select id="selectElement" multiple title="selectElement">
    <option SELECTED>Option 1</option>
    <option>Option 2</option>
    <option DISABLED>Option 3</option>
</select>

<p>This tests that we report the correct selection-related states.</p>

<p id="console"></p>

<script>
    if (window.testRunner && window.accessibilityController) {
        shouldBeTrue('accessibilityController.accessibleElementById("selectElement").isMultiSelectable');
        shouldBeTrue('accessibilityController.accessibleElementById("selectElement").childAtIndex(0).isSelectable');
        shouldBeTrue('accessibilityController.accessibleElementById("selectElement").childAtIndex(0).isSelected');
        shouldBeTrue('accessibilityController.accessibleElementById("selectElement").childAtIndex(1).isSelectable');
        shouldBeFalse('accessibilityController.accessibleElementById("selectElement").childAtIndex(1).isSelected');
        shouldBeFalse('accessibilityController.accessibleElementById("selectElement").childAtIndex(2).isSelectable');
        shouldBeFalse('accessibilityController.accessibleElementById("selectElement").childAtIndex(2).isSelected');
    }
</script>
</body>
</html>
